flask/requirements.txt
biopython==1.78
BioSequences==1.1.5
Flask==1.1.2
Flask-CLI==0.4.0
Flask-WTF==1.1.1
pandas==1.4.4
numpy==1.21.5
Jinja2==2.11.3
jinja2-time==0.2.0
MarkupSafe==2.0.1
itsdangerous==2.0.1
Werkzeug==2.0.3